+++ changelog.md
@@ -1,7 +1,9 @@
-﻿## 5.0.2
+﻿## 5.1.0
 - Debug: remove console.log for polls monitoring ([#202](https://github.com/Eccleria/ewibot/pull/202))
 - Feat: refacto on activities ([#204](https://github.com/Eccleria/ewibot/pull/204))
 - Fix: typo on anonymous check during poll refresh ([#199](https://github.com/Eccleria/ewibot/pull/199))
+- Remove: twitter unused code ([#205](https://github.com/Eccleria/ewibot/pull/205))
+- Remove: useless client attributes ([#205](https://github.com/Eccleria/ewibot/pull/205))
 
 ## 5.0.1
 - Fix: botMessage missing setDefaultMemberPermission ([#184](https://github.com/Eccleria/ewibot/pull/184))